Added functionality for moving multiple cards
Added functions takeCardStack() and addCardStack() to gameplay.h which allow the player to take a stack of more than one card and move it to a different location of the tableau. also made minor corrections on the tableau/foundations mixup, mostly just to the temp display.
Also adjusted the logic used in addCardToPiles() for checking if a move is valid as I looked it over today and realized in general it was Very Bad.